% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alleleFrequency.R
\name{alleleFrequency}
\alias{alleleFrequency}
\title{Determines the frequency of each allele in each population}
\usage{
alleleFrequency(filePath = NULL, dataIn = NULL, popInfo = NULL,
  ploidy = 2)
}
\arguments{
\item{filePath}{full path to the vcf file}

\item{dataIn}{data.table obtained using vcfLoad function}

\item{popInfo}{dataframe with  population identifier (column1) and sample
identifier (column2)}
}
\value{
A data.table with contig info (#CHROM), SNP position (POS), Allele,
Population ID (Pop), and the count of the allele (Count), estimate of allele
freuquency in a population (Frequency), and variance of the allele frequency
(Variance).
}
\description{
Takes input vcf or data.table loaded using vcfLoad function and determines
the allelic frequency of every allele in each SNP in the provided population
information
}
\examples{
# To follow
}
