% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arp2vcf.R
\name{arp2vcf}
\alias{arp2vcf}
\title{Converts fastsimcoal2 simulated DNA data to a vcf format}
\usage{
arp2vcf(filePath = NULL, outFile = NULL, ploidy = 2)
}
\arguments{
\item{filePath}{full path and name of the arp file}

\item{outFile}{full path and name of the desired outfile file}

\item{ploidy}{ploidy of the organism. Default is 2 (diploid)}
}
\value{
A data.table with structure similar to a vcf
}
\description{
This parses an arlequin formatted output of fastsimcoal2. Then, it converts
the file into a vcf. The fastsimcoal2 simulates alleles, this function
assembles those alleles into individual using the given ploidy. Individual
genotypes are labeled by individual number and population.
}
\examples{
filePath <- "D:/OneDrive/uniFiles/coalescentSimu/snputility/simuRADseq_2Pop/simuRADseq/simuRADseq_1_1.arp"
ploidy <- 2
x <- arp2vcf(filePath = filePath, ploidy = ploidy)
}
