% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contigHap.R
\name{contigHap}
\alias{contigHap}
\title{Assembles the contig haplotype from a phased vcf}
\usage{
contigHap(filePath = NULL, dataIn)
}
\arguments{
\item{filePath}{full path to the vcf file; this calls the vcfLoad}

\item{dataIn}{data.table obtained using vcfLoad function}
}
\value{
A data.table with the contig name, sample id and haplotype.
}
\description{
Merge SNPs from the same contig to generate haplotype data; the vcf
file must be phased.
}
\examples{
# To follow
}
