% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findfixedSNP.R
\name{findfixedSNP}
\alias{findfixedSNP}
\title{Calculates allele frequency in each population.}
\usage{
findfixedSNP(filePath = NULL, dataIn, popinfo)
}
\arguments{
\item{filePath}{full path to the vcf file; this calls the vcfLoad}

\item{dataIn}{data.table obtained using vcfLoad function}

\item{popinfo}{dataframe with  population identifier (column1) and sample identifier (column2)}
}
\value{
A data.table with population.allele as columns and SNPs as rows with frequencies as values
}
\description{
Determines the allele frequencies of each SNPs in specifed populations
}
\examples{
# reads example.vf, reads table containing sample id and population
# id of each sample then computes allele frequency for each snp in
# each population
vcfInput <- vcfLoad("example.vcf")
populationData <- read.table("populationMetadata")
findfixedSNP(dataIn = vcfInput, popinfo = populationData)
}
