% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlleleCount.R
\name{getAlleleCount}
\alias{getAlleleCount}
\title{Determines the number of copies of each allele of a locus in every sample}
\usage{
getAlleleCount(tempRes)
}
\arguments{
\item{tempRes}{data.table summarizing the alleles present in all samples.
This should be an output of recodeGT function.}
}
\value{
A data.table with contig info (#CHROM), SNP position (POS), Allele,
sample ID, and the count of the allele of a particular SNP in an individual
}
\description{
Counts the number of each allele of a locus in each individual. If an individual
or sample is not genotyped for a certain locus, the count will be marked as NA.
}
\examples{
# To follow
}
