% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeReads.R
\name{mergeReads}
\alias{mergeReads}
\title{Reads raw fastq files and concatenates the forward and reverse read}
\usage{
mergeReads(filePath, fastqNam, outFile = NULL, chunkLines = 2e+05)
}
\arguments{
\item{filePath}{full path to the directory containing the forward and reverse reads}

\item{fastqNam}{name of the fastq file. See Details}

\item{outFile}{name of the fastq file to be generated}
}
\value{
NULL; saves a fastq file instead.
}
\description{
Merge the forward and reverse reads of fastq files. This basically concatenates
the sequence and the read quality and uses the sequence infromation and other
metadata of the forward read. The fastqfile should have the following name format:
name.F.fq for the forward and name.R.fq for the reverse.
}
\examples{
mergeReads(filePath = "rawReads/", fasqNam = "EMSE_5102")
}
