% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeGT.R
\name{recodeGT}
\alias{recodeGT}
\title{Recode the genotype with the allele}
\usage{
recodeGT(dataIn)
}
\arguments{
\item{dataIn}{data.table obtained using vcfLoad function}
}
\value{
A data.table with contig info (#CHROM), SNP position (POS), sample ID
and the alleles of the sample for that particular SNP
}
\description{
Converts the genotype written by vcf using numbers to allele. This is
internally called by other functions that are involved in getting allele
counts. Moreover, this functions simplifies the data by removing other
information of the vcf.
}
\examples{
# To follow
}
