% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2genind.R
\name{vcf2genind}
\alias{vcf2genind}
\title{Fast conversion of vcf to genind}
\usage{
vcf2genind(filePath = NULL, inputData = NULL, pop = NULL,
  prevcall = NULL, ploidy = NULL, type = "codom", strata = NULL,
  hierarchy = NULL)
}
\arguments{
\item{filePath}{full path to the vcf file; this calls the vcfLoad}

\item{pop}{vector of the the popID of each sample}

\item{dataIn}{data.table obtained using vcfLoad function}

\item{...}{other parameters needed by genind type object, see genind or df2genind}
}
\value{
A genind class object to be used by adegenet
}
\description{
This function converts a raw vcf file to a genind object with the
assumptions that (1) the ploidy is 2 (but will work with one actually,
but not with ploidy > 2) and (2) marker is codominant (duhhhhhh, it
should be!!!) this basically creates an allele frequency table from
the vcf which is the used as the tab in a genind object. Then, the
output is wrapped using the genind function from adegenet together
with other auxillary information.
}
\examples{
# reads the example.vcf and makes a genind object
vcfGenind <- vcf2genind("example.vcf")
populationData <- read.table("populationMetadata")
findfixedSNP(dataIn = vcfInput, popinfo = populationData)
}
