% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfLoad.R
\name{vcfLoad}
\alias{vcfLoad}
\title{Loads a vcf and parse by SNP information per sample}
\usage{
vcfLoad(filePath = NULL, SNPstart = NA, SNPend = NA)
}
\arguments{
\item{filePath}{full path to the vcf file}

\item{SNPstart}{number specifying which SNP to start to read}

\item{SNPend}{number specifying which SNP to read last}
}
\value{
A data.table of parsed SNP in long format
}
\description{
Loads a vcf and parse by SNP information per sample
}
\examples{
# reads all SNPs from the example.vcf
vcfLoad(filePath = "example.vcf")
# reads the 500th up to 3000th SNP
vcfLoad(filePath = "example.vcf", SNPstart = 500, SNPend = 3000)
}
